requirements.txt
anyio==4.12.0
asgiref==3.11.0
backports.tarfile==1.2.0
build==1.3.0
CacheControl==0.14.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cleo==2.1.0
click==8.3.1
crashtest==0.4.1
distlib==0.4.0
Django==5.2.9
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dulwich==0.24.10
fastjsonschema==2.21.2
filelock==3.20.1
findpython==0.7.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
keyring==25.7.0
more-itertools==10.8.0
msgpack==1.1.2
packaging==25.0
pbs-installer==2025.12.17
pip-tools==7.5.2
pkginfo==1.12.1.2
platformdirs==4.5.1
poetry==2.2.1
poetry-core==2.2.1
pycparser==2.23
PyJWT==2.10.1
pyproject_hooks==1.2.0
python-dotenv==1.2.1
RapidFuzz==3.14.3
requests==2.32.5
requests-toolbelt==1.0.0
shellingham==1.5.4
sqlparse==0.5.4
tomlkit==0.13.3
trove-classifiers==2025.12.1.14
typing_extensions==4.15.0
urllib3==2.6.2
virtualenv==20.35.4
xattr==1.3.0
zipp==3.23.0
zstandard==0.25.0
